Phase 2: Splitting up the files to match the names
In the previous step the code was divided inside files and the tests
split out.

They have now been moved to files which match the names they have
and the tests have been moved about and coalesced appropriately.
