+++ tests/unit/lms/views/lti/__init___test.py
@@ -5,22 +5,22 @@ import pytest
 from lms.services import HAPIError
 from lms.services.h_api_client import HAPIClient
 from lms.values import HUser
-from lms.views.basic_lti_launch import BasicLTILaunchViews
+from lms.views.lti import LTIViewBaseClass
 
 
-class TestBasicLTILaunch:
+class TestLTIViewBaseClass:
     """
     Test behavior common to all LTI launches.
     """
 
     def test_it_configures_frontend(self, context, pyramid_request):
-        BasicLTILaunchViews(context, pyramid_request)
+        LTIViewBaseClass(context, pyramid_request)
         assert context.js_config["mode"] == "basic-lti-launch"
 
     def test_it_does_not_configure_grading_if_request_unqualified(
         self, context, pyramid_request
     ):
-        BasicLTILaunchViews(context, pyramid_request)
+        LTIViewBaseClass(context, pyramid_request)
         assert "lmsGrader" not in context.js_config
 
     def test_it_adds_report_submission_config_if_required_params_present(
@@ -28,7 +28,7 @@ class TestBasicLTILaunch:
     ):
         pyramid_request.params.update(lti_outcome_params)
 
-        BasicLTILaunchViews(context, pyramid_request)
+        LTIViewBaseClass(context, pyramid_request)
 
         assert context.js_config["submissionParams"] == {
             "h_username": context.h_user.username,
@@ -50,7 +50,7 @@ class TestBasicLTILaunch:
         pyramid_request.params.update(lti_outcome_params)
         del pyramid_request.params[key]
 
-        BasicLTILaunchViews(context, pyramid_request)
+        LTIViewBaseClass(context, pyramid_request)
 
         assert "submissionParams" not in context.js_config
 
@@ -62,7 +62,7 @@ class TestBasicLTILaunch:
             {"tool_consumer_info_product_family_code": "whiteboard"}
         )
 
-        BasicLTILaunchViews(context, pyramid_request)
+        LTIViewBaseClass(context, pyramid_request)
 
         assert "submissionParams" not in context.js_config
 
@@ -75,7 +75,7 @@ class TestBasicLTILaunch:
             authority="TEST_AUTHORITY", username="user123", display_name="Jim Smith"
         )
 
-        BasicLTILaunchViews(context, pyramid_request)
+        LTIViewBaseClass(context, pyramid_request)
 
         h_api_client.get_user.assert_called_once_with("user123")
         assert context.hypothesis_config["focus"] == {
@@ -89,7 +89,7 @@ class TestBasicLTILaunch:
         pyramid_request.params.update({"focused_user": "user123"})
         h_api_client.get_user.side_effect = HAPIError("User does not exist")
 
-        BasicLTILaunchViews(context, pyramid_request)
+        LTIViewBaseClass(context, pyramid_request)
 
         h_api_client.get_user.assert_called_once_with("user123")
         assert context.hypothesis_config["focus"] == {
